+++ src/Requests/Exceptions/ConflictRequestException.php
@@ -1,0 +1,5 @@
+<?php namespace Ordercloud\Requests\Exceptions;
+
+class ConflictRequestException extends OrdercloudRequestException
+{
+}