+++ src/Requests/Products/DeleteProductTagRequest.php
@@ -1,0 +1,27 @@
+<?php namespace Ordercloud\Requests\Products;
+
+use Ordercloud\Support\CommandBus\Command;
+
+class DeleteProductTagRequest implements Command
+{
+    /**
+     * @var int
+     */
+    private $tagId;
+
+    /**
+     * @param int $tagId
+     */
+    public function __construct($tagId)
+    {
+        $this->tagId = $tagId;
+    }
+
+    /**
+     * @return int
+     */
+    public function getTagId()
+    {
+        return $this->tagId;
+    }
+}