+++ changelog.md
@@ -2,6 +2,19 @@
 All notable changes to this project will be documented here.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.5.0]
+### Added
+- `ConflictRequestException`
+- `AbstractDeleteRequestHandler`
+- `ProductTagService::deleteTag()`
+- `ConnectionService`
+
+### Depricated
+- `OrganisationService::getConnectionsByType()`
+- `OrganisationService::getChildConnections()`
+- `OrganisationService::getMarketplaceConnections()`
+- `OrganisationService::getConnections()`
+
 ## [1.4.0]
 ### Added
 - `ProductTagService`, `ProductTagCollection` & `TagCriteria`
@@ -140,7 +153,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [< 0.3.1]
 ### History unavailable
 
-[Unreleased]: https://github.com/ordercloud/ordercloud-php/compare/1.4.0...HEAD
+[Unreleased]: https://github.com/ordercloud/ordercloud-php/compare/1.5.0...HEAD
+[1.5.0]: https://github.com/ordercloud/ordercloud-php/compare/1.4.0...1.5.0
 [1.4.0]: https://github.com/ordercloud/ordercloud-php/compare/1.3.1...1.4.0
 [1.3.1]: https://github.com/ordercloud/ordercloud-php/compare/1.3.0...1.3.1
 [1.3.0]: https://github.com/ordercloud/ordercloud-php/compare/1.2.0...1.3.0